Bone Fix Refactor
Updated to fix an issue where the builder would expect additional bones to be in a specific place but they weren't. This does create more places where other things need to be specific, but these should be easier for users to fix.

TODO: add more warnings for areas where things could go wrong. Like names not being the same.